requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

fastapi
uvicorn
gradio
pyvi
rank-bm25
qdrant-client
sentence-transformers
pypdf
pdfplumber
pymupdf
pydantic
pydantic_settings
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
tqdm
transformers
bitsandbytes
accelerate
google-generativeai
